+++ lib/yt/models/channel.rb
@@ -32,6 +32,12 @@ module Yt
       has_report :shares
 
       # @macro has_report
+      has_report :subscribers_gained
+
+      # @macro has_report
+      has_report :subscribers_lost
+
+      # @macro has_report
       has_report :impressions
 
       # @macro has_viewer_percentages